requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
importlib-metadata==4.11.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
protobuf==3.19.4
python-dateutil==1.5
requests==2.27.1
SQLAlchemy==1.4.32
tk==0.1.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.7.0
